backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
openai==1.3.5
python-multipart==0.0.6
pydantic==2.5.0
PyPDF2==3.0.1
python-dotenv==1.0.0
httpx==0.27.2